レビュー指摘対応: 読み込み競合とサイズ検証のオーバーフローを修正
- 画像を短時間に連続選択した際、先に開始した読み込みが後から完了して
  新しい選択結果を上書きしないよう、進行中の読み込みジョブをキャンセルする。
  キャンセルはエラー表示せず再送出する
- BMP / GIF エンコーダーの pixels サイズ検証で width * height を Long で
  計算し、Int オーバーフローによる誤判定を防ぐ (65536x65536 が 0 に
  ラップするケースのテストを追加)

+++ app/src/main/java/jp/developer/bbee/featuredemo/image/BmpEncoder.kt
@@ -19,7 +19,8 @@ object BmpEncoder {
 
     fun encode(pixels: IntArray, width: Int, height: Int): ByteArray {
         require(width > 0 && height > 0) { "width と height は正の値である必要があります" }
-        require(pixels.size == width * height) { "pixels のサイズが width * height と一致しません" }
+        // width * height は Int でオーバーフローし得るため Long で比較する
+        require(pixels.size.toLong() == width.toLong() * height) { "pixels のサイズが width * height と一致しません" }
 
         val rowBytes = width * 3
         // 各行は 4 バイト境界にパディングされる
