+++ app/src/main/java/jp/developer/bbee/featuredemo/image/GifEncoder.kt
@@ -22,7 +22,8 @@ object GifEncoder {
     fun encode(pixels: IntArray, width: Int, height: Int): ByteArray {
         require(width > 0 && height > 0) { "width と height は正の値である必要があります" }
         require(width <= 0xFFFF && height <= 0xFFFF) { "GIF の最大サイズ (65535px) を超えています" }
-        require(pixels.size == width * height) { "pixels のサイズが width * height と一致しません" }
+        // width * height は Int でオーバーフローし得るため Long で比較する
+        require(pixels.size.toLong() == width.toLong() * height) { "pixels のサイズが width * height と一致しません" }
 
         val out = ByteArrayOutputStream(pixels.size / 2 + 1024)
